+++ src/resources/active/v1/accounts/index.ts
@@ -40,9 +40,9 @@ export {
 export {
   Orders,
   type ApStrategy,
+  type APIDecimal64,
   type BaseStrategyParams,
   type DarkStrategy,
-  type Destination,
   type DmaStrategy,
   type OrderStatus,
   type OrderStrategy,