+++ src/resources/active/v1/instruments/index.ts
@@ -11,10 +11,14 @@ export {
 } from './analyst-reporting';
 export {
   Events,
+  type InstrumentAllEventsData,
   type InstrumentDividendEvent,
+  type InstrumentEventsByDate,
   type InstrumentEventsData,
   type InstrumentSplitEvent,
+  type EventGetAllInstrumentEventsResponse,
   type EventGetInstrumentEventsResponse,
+  type EventGetAllInstrumentEventsParams,
   type EventGetInstrumentEventsParams,
 } from './events';
 export {
@@ -24,6 +28,7 @@ export {
   type InstrumentCoreList,
   type InstrumentEarnings,
   type InstrumentQuote,
+  type InstrumentSecurityID,
   type InstrumentGetInstrumentByIDResponse,
   type InstrumentGetInstrumentsResponse,
   type InstrumentGetInstrumentByIDParams,