+++ src/resources/active/v1/instruments/instruments.ts
@@ -15,10 +15,14 @@ import {
 } from './analyst-reporting';
 import * as EventsAPI from './events';
 import {
+  EventGetAllInstrumentEventsParams,
+  EventGetAllInstrumentEventsResponse,
   EventGetInstrumentEventsParams,
   EventGetInstrumentEventsResponse,
   Events,
+  InstrumentAllEventsData,
   InstrumentDividendEvent,
+  InstrumentEventsByDate,
   InstrumentEventsData,
   InstrumentSplitEvent,
 } from './events';
@@ -161,11 +165,6 @@ export interface Instrument extends InstrumentCore {
   long_concentration_limit?: string | null;
 
   /**
-   * The percent of a long position's value you must post as margin
-   */
-  long_margin_rate?: string | null;
-
-  /**
    * The total market capitalization
    */
   market_cap?: string | null;
@@ -195,18 +194,14 @@ export interface Instrument extends InstrumentCore {
    * short side
    */
   short_concentration_limit?: string | null;
-
-  /**
-   * The percent of a short position's value you must post as margin
-   */
-  short_margin_rate?: string | null;
 }
 
-/**
- * Represents a tradable financial instrument, as a more concise item listing only
- * key fields.
- */
 export interface InstrumentCore {
+  /**
+   * Unique instrument identifier
+   */
+  id: string;
+
   /**
    * The ISO country code of the instrument's issue
    */
@@ -248,14 +243,23 @@ export interface InstrumentCore {
   is_threshold_security: boolean;
 
   /**
-   * A unique Clear Street identifier for the instrument
+   * @deprecated Deprecated. Use `security_ids`.
+   *
+   * A primary security identifier for this instrument.
    */
   security_id: string;
 
   /**
-   * The source system for the security identifier
+   * Deprecated. Use `security_ids`.
+   *
+   * The source for `security_id`.
    */
   security_id_source: V1API.SecurityIDSource;
+
+  /**
+   * All known security identifiers for this instrument
+   */
+  security_ids: Array<InstrumentSecurityID>;
 
   /**
    * The trading symbol for the instrument
@@ -273,6 +277,11 @@ export interface InstrumentCore {
   expiry?: string | null;
 
   /**
+   * The percent of a long position's value you must post as margin
+   */
+  long_margin_rate?: string | null;
+
+  /**
    * The full name of the instrument or its issuer
    */
   name?: string | null;
@@ -281,6 +290,11 @@ export interface InstrumentCore {
    * The type of security (e.g., Common Stock, ETF)
    */
   security_type?: V1API.SecurityType | null;
+
+  /**
+   * The percent of a short position's value you must post as margin
+   */
+  short_margin_rate?: string | null;
 
   /**
    * The strike price for options instruments
@@ -358,6 +372,22 @@ export interface InstrumentQuote {
    * The total number of shares traded during the current trading day
    */
   volume: number;
+}
+
+/**
+ * Represents a tradable financial instrument, as a more concise item listing only
+ * key fields.
+ */
+export interface InstrumentSecurityID {
+  /**
+   * The identifier for the instrument
+   */
+  security_id: string;
+
+  /**
+   * The source system for the security identifier
+   */
+  security_id_source: V1API.SecurityIDSource;
 }
 
 export interface InstrumentGetInstrumentByIDResponse extends Shared.BaseResponse {
@@ -474,6 +504,7 @@ export declare namespace Instruments {
     type InstrumentCoreList as InstrumentCoreList,
     type InstrumentEarnings as InstrumentEarnings,
     type InstrumentQuote as InstrumentQuote,
+    type InstrumentSecurityID as InstrumentSecurityID,
     type InstrumentGetInstrumentByIDResponse as InstrumentGetInstrumentByIDResponse,
     type InstrumentGetInstrumentsResponse as InstrumentGetInstrumentsResponse,
     type InstrumentGetInstrumentByIDParams as InstrumentGetInstrumentByIDParams,
@@ -492,10 +523,14 @@ export declare namespace Instruments {
 
   export {
     Events as Events,
+    type InstrumentAllEventsData as InstrumentAllEventsData,
     type InstrumentDividendEvent as InstrumentDividendEvent,
+    type InstrumentEventsByDate as InstrumentEventsByDate,
     type InstrumentEventsData as InstrumentEventsData,
     type InstrumentSplitEvent as InstrumentSplitEvent,
+    type EventGetAllInstrumentEventsResponse as EventGetAllInstrumentEventsResponse,
     type EventGetInstrumentEventsResponse as EventGetInstrumentEventsResponse,
+    type EventGetAllInstrumentEventsParams as EventGetAllInstrumentEventsParams,
     type EventGetInstrumentEventsParams as EventGetInstrumentEventsParams,
   };