+++ src/resources/active/v1/watchlists/index.ts
@@ -1,0 +1,20 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+export {
+  Items,
+  type AddWatchlistItemData,
+  type ItemAddWatchlistItemResponse,
+  type ItemAddWatchlistItemParams,
+  type ItemDeleteWatchlistItemParams,
+} from './items';
+export {
+  Watchlists,
+  type WatchlistDetail,
+  type WatchlistEntry,
+  type WatchlistEntryList,
+  type WatchlistItemEntry,
+  type WatchlistCreateWatchlistResponse,
+  type WatchlistGetWatchlistByIDResponse,
+  type WatchlistGetWatchlistsResponse,
+  type WatchlistCreateWatchlistParams,
+} from './watchlists';